requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
openai==1.74.0
python-dotenv==1.0.0
jinja2==3.1.2
watchdog==3.0.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
alembic==1.12.0
pydantic[email]==2.2.1 